% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{distributioninfo}
\alias{distributioninfo}
\title{A table with geographic distribution to run the vignette}
\format{
A data frame where each species has a distribution associated
}
\description{
An example of trait information in the right format for lemad
}
