% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_times.R
\name{event_times}
\alias{event_times}
\title{Event times of a (possibly non-ultrametric) phylogenetic tree}
\usage{
event_times(phy)
}
\arguments{
\item{phy}{phylogenetic tree of class phylo, without polytomies, rooted and 
with branch lengths. Need not be ultrametric.}
}
\value{
times at which speciation or extinction happens.
}
\description{
Times at which speciation or extinction occurs
}
\note{
This script has been modified from BAMMtools' internal function 
NU.branching.times
}
