% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemad_utils.R
\name{give_me_states_combination}
\alias{give_me_states_combination}
\title{All possible region combination.}
\usage{
give_me_states_combination(areas, num_max_multiregion)
}
\arguments{
\item{areas}{Vector with single region It is NOT the species presence.}

\item{num_max_multiregion}{integer indicating the maximum number of regions a lineage can possible take at a given point in time. It can go from 2 to length(areas).}
}
\value{
A vector with all the combinations of regions
}
\description{
Combines the regions into all the possible multi-region distribution.
}
\examples{
areas <- c("A", "B", "C")
give_me_states_combination(areas,num_max_multiregion = 3)
}
