% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_lemad_ml.R
\name{lemad_analysis}
\alias{lemad_analysis}
\title{Maximum likehood optimization for (lemad) and probabilities of ancestral lineages at regions}
\usage{
lemad_analysis(
  phylotree_recons,
  species_presence,
  areas,
  num_max_multiregion,
  condition_on_origin = NULL,
  DEC_events,
  missing_spp_areas,
  lineage_extinction = "free",
  initial_lambda = NULL,
  initial_disperextirpation = NULL
)
}
\arguments{
\item{phylotree_recons}{phylogenetic tree of class phylo, ultrametric, rooted and with branch lengths.}

\item{species_presence}{a vector (of class character) with regions for each species. Coded in letters: "C"  "A"  "C"  "C" . Same order than tree tips}

\item{areas}{a vector that includes all the regions. Do not include any combination of locations}

\item{num_max_multiregion}{indicates the maximum number of regions a lineage can possible take at a given point in time. It can go from 2 to length(areas). Alternatively, it can be a vector of length 2: first element to be the maximum number of regions and the second a letter that represents "without any modern species" e.g.,num_max_multiregion = c(3,"D"). See vignette}

\item{condition_on_origin}{If NULL, the conditioning on the crown origin follows Herrera-Alsina et al. 2019 (Syst. Biol.). But the user can provide the hypothesized root state: condition_on_origin <- "A"}

\item{DEC_events}{DIVA and DEC models differ sligthly in the type of events that are allowed e.g., vacariance narrow}

\item{missing_spp_areas}{a list informing the proportion of missing species (missing from the tree) per location. Only include incomplete locations in this list. See example. If tree is complete, use NULL.}

\item{lineage_extinction}{set "free" to have it estimated or provide a rate to fix it at.}

\item{initial_lambda}{Vector of length 2 for initial in-situ and vicariance rates to start the ML search. If NULL, the starting lambda will be taken from a Birth-death process.}

\item{initial_disperextirpation}{Intial values for dispersal-local extinction during the ML search. If NULL, it will be equal to lambda/5. For advanced analysis, it can take a transition matrix to allow or restrict certain dispersal or extirpation events. See vignette}
}
\value{
List with model's loglik, number of free parameters, estimates of rates, and ancestral locations probabilities.
}
\description{
Maximum likehood search under Lineage Extinction Model of Ancestral Distribution (LEMAD)
}
\examples{
# Example of how to set the arguments for a Maximum Likelihood search.
library(lemad)
library(DDD)
areas <- c("A", "B", "C")
set.seed(4)
phylotree_recons <- ape::rcoal(4, tip.label = 1:4)
species_presence <- c("B","C","A","AC")
DEC_events <- FALSE
# In case the tree has  80\% of species inhabiting in A. And 95\% of species in AC.
# Notice that regions in vector 1 and proportions in vector 2 have to be in the same order
missing_spp_areas <- list()
missing_spp_areas[[1]] <- c("A","AC")
missing_spp_areas[[2]] <- c(0.8,0.95)

output <- lemad_analysis(
phylotree_recons,
species_presence,
areas,
num_max_multiregion = 3,
condition_on_origin = NULL,
DEC_events,
missing_spp_areas,
lineage_extinction =  0.005,
initial_lambda = NULL,
initial_disperextirpation = NULL
)

output$model_ml #  -9.893469 the loglikelihood for the model
}
