% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemad_utils.R
\name{lemad_prepare_q_matrix}
\alias{lemad_prepare_q_matrix}
\title{Transition matrix preparation.}
\usage{
lemad_prepare_q_matrix(
  all_area_combination,
  matrices_names,
  id_q_expansion,
  id_q_contraction
)
}
\arguments{
\item{all_area_combination}{Matrix with all regions combinations coming from function prepare_full_lambdas_vicariance().}

\item{matrices_names}{Vector with all regions names coming from function prepare_full_lambdas_vicariance().}

\item{q_expansion}{Parameter ID for colonization rate.}

\item{q_contraction}{Parameter ID for contraction (local extinction/extirpation) rate.}
}
\value{
the transtion matrix.
}
\description{
Prepares a q matrix for range expansion and contraction.
}
