% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{phylo_Vign2}
\alias{phylo_Vign2}
\title{A phylogenetic reconstuction to run the vignette}
\format{
Phylogenetic tree in format nexus, rooted, including branch lengths
}
\description{
An example phylogeny in the right format for lemad
}
