% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemad_utils.R
\name{prepare_full_lambdas_vicariance}
\alias{prepare_full_lambdas_vicariance}
\title{Making lambda matrices.}
\usage{
prepare_full_lambdas_vicariance(
  areas,
  num_max_multiregion,
  id_rate_insitu,
  id_rate_vicariance,
  DEC_events
)
}
\arguments{
\item{areas}{Vector with single regions. It is NOT the species presence.}

\item{id_rate_insitu}{Parameter ID for the rate of insitu speciation.}

\item{id_rate_vicariance}{Parameter ID for the rate of vicariance}
}
\value{
A list of matrices but also the names of the states (the combinations of all locations).
}
\description{
Prepares the lambda (speciation) matrices .
}
\examples{
all_locations <- c("A","B","C") 
id_rate_insitu <- 1
id_rate_vicariance <- 2
num_max_multiregion <- 3
DEC_events <- FALSE
prepare_full_lambdas_vicariance(all_locations,num_max_multiregion,id_rate_insitu,id_rate_vicariance,DEC_events)
}
